add SignUpTitleTextWidget and  FormSignUp
