+++ lib/feature/auth/register/presentation/widgets/sign_up_title_text_widget.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/widgets.dart';
+
+class SignUpTitleTextWidget extends StatelessWidget {
+  const SignUpTitleTextWidget({
+    super.key,
+    required this.title,
+    required this.style,
+  });
+  final String title;
+  final TextStyle style;
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(child: Text(title, style: style));
+  }
+}
